requirements.txt
Django>=2
django-extensions
djangorestframework
markdown
django-filter
django-crispy-forms == 1.9.0
psycopg2-binary
django-cors-headers